Ensure duplicate provider users aren't orphaned
If there are provider users who are solely associated with the duplicate provider, do not destroy the provider record.
We should address any cases like this in support. No users currently are in this state.

+++ app/services/data_migrations/remove_duplicate_provider.rb
@@ -12,16 +12,23 @@ module DataMigrations
       if permissions.any?
         permissions.each do |permission|
           ActiveRecord::Base.transaction do
-            courses = permission.training_provider.courses.where(accredited_provider: permission.ratifying_provider)
-
             next if permission.ratifying_provider.courses.any?
+
+            courses = permission.training_provider.courses.where(accredited_provider: permission.ratifying_provider)
 
             courses.each do |course|
               course.update!(accredited_provider: nil, audit_comment: 'Deduplicating accredited provider, this course is self-ratified')
             end
 
             permission.destroy!
-            permission.ratifying_provider.destroy!
+
+            # Ensure we don't orphan any users who were only associated with the duplicate provider
+            if permission.ratifying_provider.provider_users.select { |user| user.providers.size == 1 }.any?
+              Rails.logger.warn "Skipping deletion of #{permission.ratifying_provider.name}. This organisation has users which do not belong to any other provider."
+            else
+              permission.destroy!
+              permission.ratifying_provider.destroy!
+            end
           end
         end
       end
